feat (Table, TableGroup): Table, TableGroup 모듈 분리
