+++ module-table/src/main/java/kitchenpos/table/application/TableService.java
@@ -1,0 +1,55 @@
+package kitchenpos.table.application;
+
+import kitchenpos.table.dto.EmptyRequest;
+import kitchenpos.table.dto.NumberOfGuestsRequest;
+import kitchenpos.table.dto.OrderTableRequest;
+import kitchenpos.table.dto.OrderTableResponse;
+import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.domain.OrderTableRepository;
+import kitchenpos.table.exception.NotFoundOrderTableException;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+
+@Service
+@Transactional(readOnly = true)
+public class TableService {
+    private final OrderTableRepository orderTableRepository;
+
+    public TableService(OrderTableRepository orderTableRepository) {
+        this.orderTableRepository = orderTableRepository;
+    }
+
+    @Transactional
+    public OrderTableResponse create(OrderTableRequest request) {
+        return OrderTableResponse.of(orderTableRepository.save(request.toOrderTable()));
+    }
+
+    public List<OrderTableResponse> list() {
+        return OrderTableResponse.of(orderTableRepository.findAll());
+    }
+
+    @Transactional
+    public OrderTableResponse changeEmpty(Long orderTableId, EmptyRequest request) {
+        OrderTable orderTable = findById(orderTableId);
+        orderTable.changeEmpty(request.isEmpty());
+        return OrderTableResponse.of(orderTableRepository.save(orderTable));
+    }
+
+    @Transactional
+    public OrderTableResponse changeNumberOfGuests(Long orderTableId, NumberOfGuestsRequest request) {
+        OrderTable orderTable = findById(orderTableId);
+        orderTable.changeNumberOfGuests(request.getNumberOfGuests());
+        return OrderTableResponse.of(orderTableRepository.save(orderTable));
+    }
+
+    public List<OrderTable> findAllIdIn(List<Long> ids) {
+        return orderTableRepository.findAllByIdIn(ids);
+    }
+
+    private OrderTable findById(Long id) {
+        return orderTableRepository.findById(id)
+                                   .orElseThrow(NotFoundOrderTableException::new);
+    }
+}
