+++ module-table/src/main/java/kitchenpos/table/domain/TableEmptyChangedEvent.java
@@ -1,0 +1,14 @@
+package kitchenpos.table.domain;
+
+public class TableEmptyChangedEvent {
+
+    private final Long orderTableId;
+
+    public TableEmptyChangedEvent(Long orderTableId) {
+        this.orderTableId = orderTableId;
+    }
+
+    public Long getOrderTableId() {
+        return orderTableId;
+    }
+}
