+++ module-table/src/main/java/kitchenpos/table/exception/CannotChangeEmptyException.java
@@ -1,0 +1,8 @@
+package kitchenpos.table.exception;
+
+public class CannotChangeEmptyException extends RuntimeException {
+
+    public CannotChangeEmptyException() {
+        super("비움 여부를 변경할 수 없습니다.");
+    }
+}
