+++ module-table/src/main/java/kitchenpos/table/exception/CannotChangeNumberOfGuestsException.java
@@ -1,0 +1,8 @@
+package kitchenpos.table.exception;
+
+public class CannotChangeNumberOfGuestsException extends RuntimeException {
+
+    public CannotChangeNumberOfGuestsException() {
+        super("손님 수를 변경할 수 없습니다.");
+    }
+}
