+++ module-table/src/main/java/kitchenpos/table/exception/InvalidOrderTablesException.java
@@ -1,0 +1,8 @@
+package kitchenpos.table.exception;
+
+public class InvalidOrderTablesException extends RuntimeException {
+
+    public InvalidOrderTablesException(String message) {
+        super(message);
+    }
+}
