+++ module-table/src/main/java/kitchenpos/table/exception/NotFoundOrderTableException.java
@@ -1,0 +1,8 @@
+package kitchenpos.table.exception;
+
+public class NotFoundOrderTableException extends RuntimeException {
+
+    public NotFoundOrderTableException() {
+        super("테이블을 찾을 수 없습니다.");
+    }
+}
