+++ module-table/src/main/java/kitchenpos/table/ui/TableRestController.java
@@ -1,0 +1,53 @@
+package kitchenpos.table.ui;
+
+import kitchenpos.table.application.TableService;
+import kitchenpos.table.dto.EmptyRequest;
+import kitchenpos.table.dto.NumberOfGuestsRequest;
+import kitchenpos.table.dto.OrderTableRequest;
+import kitchenpos.table.dto.OrderTableResponse;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.net.URI;
+import java.util.List;
+
+@RestController
+public class TableRestController {
+    private final TableService tableService;
+
+    public TableRestController(final TableService tableService) {
+        this.tableService = tableService;
+    }
+
+    @PostMapping("/api/tables")
+    public ResponseEntity<OrderTableResponse> create(@RequestBody OrderTableRequest request) {
+        OrderTableResponse created = tableService.create(request);
+        URI uri = URI.create("/api/tables/" + created.getId());
+        return ResponseEntity.created(uri)
+                .body(created);
+    }
+
+    @GetMapping("/api/tables")
+    public ResponseEntity<List<OrderTableResponse>> list() {
+        return ResponseEntity.ok()
+                .body(tableService.list());
+    }
+
+    @PutMapping("/api/tables/{orderTableId}/empty")
+    public ResponseEntity<OrderTableResponse> changeEmpty(
+            @PathVariable Long orderTableId,
+            @RequestBody EmptyRequest request
+    ) {
+        return ResponseEntity.ok()
+                .body(tableService.changeEmpty(orderTableId, request));
+    }
+
+    @PutMapping("/api/tables/{orderTableId}/number-of-guests")
+    public ResponseEntity<OrderTableResponse> changeNumberOfGuests(
+            @PathVariable Long orderTableId,
+            @RequestBody NumberOfGuestsRequest request
+    ) {
+        return ResponseEntity.ok()
+                .body(tableService.changeNumberOfGuests(orderTableId, request));
+    }
+}
