+++ module-tablegroup/src/main/java/kitchenpos/tablegroup/dto/OrderTableIdRequest.java
@@ -1,0 +1,17 @@
+package kitchenpos.tablegroup.dto;
+
+public class OrderTableIdRequest {
+
+    private Long id;
+
+    public OrderTableIdRequest(Long id) {
+        this.id = id;
+    }
+
+    public OrderTableIdRequest() {
+    }
+
+    public Long getId() {
+        return id;
+    }
+}
