+++ module-tablegroup/src/main/java/kitchenpos/tablegroup/exception/InvalidTableGroupException.java
@@ -1,0 +1,8 @@
+package kitchenpos.tablegroup.exception;
+
+public class InvalidTableGroupException extends RuntimeException {
+
+    public InvalidTableGroupException(String message) {
+        super(message);
+    }
+}
