+++ module-tablegroup/src/main/java/kitchenpos/tablegroup/ui/TableGroupRestController.java
@@ -1,0 +1,33 @@
+package kitchenpos.tablegroup.ui;
+
+import kitchenpos.tablegroup.application.TableGroupService;
+import kitchenpos.tablegroup.dto.TableGroupRequest;
+import kitchenpos.tablegroup.dto.TableGroupResponse;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.net.URI;
+
+@RestController
+public class TableGroupRestController {
+    private final TableGroupService tableGroupService;
+
+    public TableGroupRestController(final TableGroupService tableGroupService) {
+        this.tableGroupService = tableGroupService;
+    }
+
+    @PostMapping("/api/table-groups")
+    public ResponseEntity<TableGroupResponse> create(@RequestBody TableGroupRequest request) {
+        TableGroupResponse response = tableGroupService.create(request);
+        URI uri = URI.create("/api/table-groups/" + response.getId());
+        return ResponseEntity.created(uri)
+                .body(response);
+    }
+
+    @DeleteMapping("/api/table-groups/{tableGroupId}")
+    public ResponseEntity<Void> ungroup(@PathVariable Long tableGroupId) {
+        tableGroupService.ungroup(tableGroupId);
+        return ResponseEntity.noContent()
+                .build();
+    }
+}
