+++ module-tablegroup/src/testFixtures/java/kitchenpos/tablegroup/domain/FixtureTableGroup.java
@@ -1,0 +1,10 @@
+package kitchenpos.tablegroup.domain;
+
+import kitchenpos.tablegroup.domain.TableGroup;
+
+public class FixtureTableGroup extends TableGroup {
+
+    public FixtureTableGroup() {
+        super(1L);
+    }
+}
