--- src/main/java/kitchenpos/table/domain/NumberOfGuests.java
@@ -1,48 +1,0 @@
-package kitchenpos.table.domain;
-
-import javax.persistence.Column;
-import javax.persistence.Embeddable;
-import java.util.Objects;
-
-@Embeddable
-public class NumberOfGuests implements Comparable<NumberOfGuests> {
-
-    @Column(nullable = false)
-    private int numberOfGuests;
-
-    public NumberOfGuests(int numberOfGuests) {
-        this.numberOfGuests = validate(numberOfGuests);
-    }
-
-    protected NumberOfGuests() {
-    }
-
-    private int validate(int numberOfGuests) {
-        if (numberOfGuests < 0) {
-            throw new IllegalArgumentException();
-        }
-        return numberOfGuests;
-    }
-
-    public int getValue() {
-        return numberOfGuests;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        NumberOfGuests that = (NumberOfGuests) o;
-        return compareTo(that) == 0;
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(numberOfGuests);
-    }
-
-    @Override
-    public int compareTo(NumberOfGuests o) {
-        return Integer.compare(numberOfGuests, o.numberOfGuests);
-    }
-}
