--- src/main/java/kitchenpos/table/domain/OrderTable.java
@@ -1,71 +1,0 @@
-package kitchenpos.table.domain;
-
-import javax.persistence.*;
-
-@Entity
-public class OrderTable {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne
-    @JoinColumn(foreignKey = @ForeignKey(name = "order_table_to_table_group"))
-    private TableGroup tableGroup;
-
-    @Embedded
-    private NumberOfGuests numberOfGuests;
-
-    @Column(nullable = false)
-    private boolean empty;
-
-    public OrderTable(int numberOfGuests, boolean empty) {
-        this.numberOfGuests = new NumberOfGuests(numberOfGuests);
-        this.empty = empty;
-    }
-
-    protected OrderTable() {
-    }
-
-    public void bindTo(TableGroup tableGroup) {
-        this.tableGroup = tableGroup;
-        this.empty = false;
-    }
-
-    public void unbind() {
-        tableGroup = null;
-    }
-
-    public boolean isGrouped() {
-        return tableGroup != null;
-    }
-
-    public boolean isEmpty() {
-        return empty;
-    }
-
-    public void changeEmpty(boolean empty) {
-        if (isGrouped()) {
-            throw new IllegalArgumentException();
-        }
-        this.empty = empty;
-    }
-
-    public void changeNumberOfGuests(int numberOfGuests) {
-        if (empty) {
-            throw new IllegalArgumentException();
-        }
-        this.numberOfGuests = new NumberOfGuests(numberOfGuests);
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public TableGroup getTableGroup() {
-        return tableGroup;
-    }
-
-    public NumberOfGuests getNumberOfGuests() {
-        return numberOfGuests;
-    }
-}
