--- src/main/java/kitchenpos/table/domain/OrderTableRepository.java
@@ -1,11 +1,0 @@
-package kitchenpos.table.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface OrderTableRepository extends JpaRepository<OrderTable, Long> {
-    List<OrderTable> findAllByIdIn(List<Long> ids);
-
-    List<OrderTable> findAllByTableGroupId(Long tableGroupId);
-}
