--- src/main/java/kitchenpos/table/dto/OrderTableIdRequest.java
@@ -1,17 +1,0 @@
-package kitchenpos.table.dto;
-
-public class OrderTableIdRequest {
-
-    private Long id;
-
-    public OrderTableIdRequest(Long id) {
-        this.id = id;
-    }
-
-    public OrderTableIdRequest() {
-    }
-
-    public Long getId() {
-        return id;
-    }
-}
