--- src/main/java/kitchenpos/table/ui/TableGroupRestController.java
@@ -1,34 +1,0 @@
-package kitchenpos.table.ui;
-
-
-import kitchenpos.table.application.TableGroupService;
-import kitchenpos.table.dto.TableGroupRequest;
-import kitchenpos.table.dto.TableGroupResponse;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.net.URI;
-
-@RestController
-public class TableGroupRestController {
-    private final TableGroupService tableGroupService;
-
-    public TableGroupRestController(final TableGroupService tableGroupService) {
-        this.tableGroupService = tableGroupService;
-    }
-
-    @PostMapping("/api/table-groups")
-    public ResponseEntity<TableGroupResponse> create(@RequestBody TableGroupRequest request) {
-        TableGroupResponse response = tableGroupService.create(request);
-        URI uri = URI.create("/api/table-groups/" + response.getId());
-        return ResponseEntity.created(uri)
-                .body(response);
-    }
-
-    @DeleteMapping("/api/table-groups/{tableGroupId}")
-    public ResponseEntity<Void> ungroup(@PathVariable Long tableGroupId) {
-        tableGroupService.ungroup(tableGroupId);
-        return ResponseEntity.noContent()
-                .build();
-    }
-}
